Major refactoring of the project directory structure. This was done to put everything in a pyTEM package so distribution makes more sense.
