+++ .gitignore
@@ -1,9 +1,9 @@
-/Testing_Interface/data/
-/Testing_Interface/TEMInferace Testing Log.xlsx
+/test/interface/data/
+/test/interface/TEMInferace Testing Log.xlsx
 /Testing_Interface/~$TEMInferace Testing Log.xlsx
-/Testing_Interface/Image_Shift_Calibration/Calibration_Data/
-/Testing_Interface/Image_Shift_Calibration/~$Image Shift Calibration.xlsx
-/Testing_Interface/test_images/
+/test/interface/Image_Shift_Calibration/Calibration_Data/
+/test/interface/Image_Shift_Calibration/~$Image Shift Calibration.xlsx
+/test/interface/test_images/
 
 # Distribution/packaging
 /build/
