+++ pyTEM/lib/interface/Acquisition.py
