+++ pyTEM/lib/interface/interface_mixins/BeamBlankerMixin.py
@@ -10,7 +10,7 @@ class BeamBlankerMixin:
     """
     Microscope beam blanker controls, including functions to blank and unblank the beam.
 
-    This mixin was developed in support of TEMInterface.TEMInterface, but can be included in other projects where helpful.
+    This mixin was developed in support of pyTEM.pyTEM, but can be included in other projects where helpful.
     """
     _tem: type(cc.CreateObject("TEMScripting.Instrument"))
