+++ pyTEM/lib/interface/interface_mixins/BeamShiftMixin.py
@@ -14,7 +14,7 @@ class BeamShiftMixin:
     """
     Microscope beam shift controls.
 
-    This mixin was developed in support of TEMInterface.TEMInterface, but can be included in other projects where helpful.
+    This mixin was developed in support of pyTEM.pyTEM, but can be included in other projects where helpful.
     """
     _tem: type(cc.CreateObject("TEMScripting.Instrument"))
     _beam_shift_matrix: type(np.empty(shape=(2, 2)))
