+++ pyTEM/lib/interface/interface_mixins/ImageShiftMixin.py
@@ -12,7 +12,7 @@ class ImageShiftMixin:
     """
     Microscope image shift controls.
 
-    This mixin was developed in support of TEMInterface.TEMInterface, but can be included in other projects where helpful.
+    This mixin was developed in support of pyTEM.pyTEM, but can be included in other projects where helpful.
     """
     _tem: type(cc.CreateObject("TEMScripting.Instrument"))
     _image_shift_matrix: type(np.empty(shape=(2, 2)))
