+++ pyTEM/lib/interface/interface_mixins/MagnificationMixin.py
@@ -13,7 +13,7 @@ class MagnificationMixin:
     Microscope magnification controls, including those for getting and setting the current microscope magnification and
      magnification index.
 
-    This mixin was developed in support of TEMInterface.TEMInterface, but can be included in other projects where helpful.
+    This mixin was developed in support of pyTEM.pyTEM, but can be included in other projects where helpful.
     """
     _tem: type(cc.CreateObject("TEMScripting.Instrument"))
