+++ pyTEM/lib/ued/AcquisitionProperties.py
