+++ pyTEM/lib/ued/obtain_shifts.py
@@ -25,8 +25,8 @@ def obtain_shifts(microscope, alphas, camera_name, verbose=False):
         Schaffer, Bernhard, Werner Grogger, and Gerald Kothleitner. “Automated Spatial Drift Correction for EFTEM
         Image Series.” Ultramicroscopy 102, no. 1 (December 2004): 27–36.
 
-    :param microscope: TEMInterface:
-        A TEMInterface interface to the microscope.
+    :param microscope: pyTEM:
+        A pyTEM interface to the microscope.
     :param alphas: np.array:
         An array of tilt angles at which to compute shifts.
     :param camera_name: str:
@@ -121,8 +121,8 @@ def _complete_one_sign(microscope, camera_name, this_signs_alphas, max_images_pe
 
     This function computes the shifts for just one sign, and by calling it twice we are able to compute all the shifts.
 
-    :param microscope: TEMInterface:
-        A TEMInterface interface to the microscope.
+    :param microscope: pyTEM:
+        A pyTEM interface to the microscope.
     :param camera_name: str:
         The name of the camera to be used for shift calculation - probably the same camera you are using for the tilt
          series itself.
@@ -228,7 +228,7 @@ if __name__ == "__main__":
     sys.path.append(str(package_directory))
 
     try:
-        from TEMInterface.TEMInterface import TEMInterface
+        from pyTEM.interface import TEMInterface
 
         scope = TEMInterface()
     except BaseException as e:
