+++ test/interface/StagePosition_testing.py
@@ -5,7 +5,7 @@ Some manual testing of the StagePosition class
 
 import sys
 sys.path.append('T:/Michael L/pyTEM')
-from TEMInterface.lib.StagePosition import StagePosition
+from pyTEM.lib.interface.StagePosition import StagePosition
 
 
 print("\nCreating a stage position object with x=0, y=1, z=2, alpha=-0.35, and beta=0.35...")
