--- uED/__init__.py
@@ -1,3 +1,0 @@
-
-__version__ = "0.1.0"
-__author__ = "BASF TEM Laboratory"
